fixing and refactoring
