+++ K6/api/building-blocks/dialogporten/serviceowner/create-dialog.js
@@ -2,8 +2,13 @@ import { check } from "k6";
 import { ServiceOwnerApiClient } from "../../../../clients/dialogporten/serviceowner/index.js";
 
 /**
+ * Function to create a dialog for a party
  * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- 
+ * @param {string} partyId - either a pid/ssn (11 digits) or a organization number (9 digits)
+ * @param {string} serviceResource - the service resource for the dialog
+ * @param {string} serviceOwner - the service owner for the dialog. an organization nunber (9 digits)
+ * @param {string} label - label for the request
+ * @return response body of the request
  */
 export function CreateDialog(
     serviceOwnerApiClient,
@@ -33,6 +38,13 @@ export function CreateDialog(
     return res.body;
 }
 
+/**
+ * Create a transmission for a dialog
+ * @param {ServiceOwnerApiClient} serviceOwnerApiClient 
+ * @param {uuidv7} dialogId - the id of the dialog to create the transmission for
+ * @param {string} label 
+ * @returns response body of the request
+ */
 export function CreateTransmission(
     serviceOwnerApiClient,
     dialogId,
@@ -55,6 +67,13 @@ export function CreateTransmission(
     return res.body;
 }
 
+/**
+ * Create an activity for a dialog
+ * @param {ServiceOwnerApiClient} serviceOwnerApiClient 
+ * @param {uuidv7} dialogId - the id of the dialog to create the activity for
+ * @param {string} label - label for the request
+ * @returns 
+ */
 export function CreateActivity(
     serviceOwnerApiClient,
     dialogId,
