+++ K6/api/tests/authentication/access-management/bff/instance-delegation-org-user.js
@@ -1,10 +1,6 @@
-import http from "k6/http";
 import exec from "k6/execution";
 import { group } from "k6";
-import { EnterpriseTokenGenerator, PersonalTokenGenerator } from "../../../../../common-imports.js";
 import { CreateDialog } from "../../../../building-blocks/dialogporten/serviceowner/index.js";
-import { ServiceOwnerApiClient } from "../../../../../clients/dialogporten/serviceowner/index.js";
-import { GraphqlClient } from "../../../../../clients/dialogporten/graphql/index.js";
 import { GetAllDialogsForPartyCheckForDialogId, GetAndVerifyDialogById } from "../../../../building-blocks/dialogporten/graphql/index.js";
 import {
     GetLookupPartyUser,
@@ -18,15 +14,12 @@ import {
     GetIsInstanceAdmin,
 } from "../../../../building-blocks/authentication/instance-delegation/index.js";
 import { GetDelegatedInstancesForResource, GetActiveConsent, GetResourceById, GetRightsMeta, DelegateRightsForResource, GetRolePermissions } from "../../../../building-blocks/authentication/client-delegations/index.js";
-import { BffUserApiClient, BffAccessManagementApiClient, BffConnectionsApiClient, BffAccessPackageApiClient } from "../../../../../clients/authentication/index.js";
 import { GetOrganizationData, CheckDelegationForResource, GetRoleMeta, GetPendingDelegationsForUser } from "../../../../building-blocks/authentication/client-delegations/index.js";
 import { GetConnections } from "../../../../building-blocks/authentication/connections/index.js";
 import { GetDelegationCheck } from "../../../../building-blocks/authentication/access-package/index.js";
-import { getItemFromList, parseCsvData, segmentData, getNumberOfVUs, getOptions } from "../../../../../helpers.js";
-import { getTokenOpts } from "./commons.js";
-
-const randomize = __ENV.RANDOMIZE ? __ENV.RANDOMIZE.toLowerCase() === "true" : false;
-
+import { getItemFromList, getOptions } from "../../../../../helpers.js";
+import { getTokenOpts, getFromTo, getClients, getDialogportenOpts, getInstanceDelegationBody } from "./commons.js";
+export { setup } from "./commons.js";
 // serviceowner which will create a dialog.
 // The yt serviceOwner is different from the other environments.
 let serviceOwnerOrgNo = "991825827";
@@ -40,17 +33,6 @@ if (__ENV.ENVIRONMENT === "yt01") {
 const resources = [
     "k6-instancedelegation-test",
 ];
-
-// All apiclient used in this test
-let serviceOwnerApiClient = undefined;
-let userApiClient = undefined;
-let accessManagementApiClient = undefined;
-let bffConnectionsApiClient = undefined;
-let bffAccessPackageApiClient = undefined;
-let graphqlClient = undefined;
-
-// personal tokengenerator.
-let personalTokenGenerator = undefined;
 
 // Use unique laber for each request to be able to check them separately in the results.
 // Number them to make sumary report easier to read and to be able to see the flow of the test in the results. 
@@ -136,47 +118,6 @@ export const options = getOptions([
 );
 
 /**
-* Function to set up and return clients to interact with the Service Owner Dialog API
-*
-* @returns {Array} An array containing the AuthorizedPartiesClient instance
-*/
-export function getClients() {
-    if (serviceOwnerApiClient == undefined) {
-        const tokenOpts = new Map();
-        tokenOpts.set("env", __ENV.ENVIRONMENT);
-        tokenOpts.set("ttl", 3600);
-        tokenOpts.set("scopes", "digdir:dialogporten.serviceprovider");
-        tokenOpts.set("org", "ttd");
-        tokenOpts.set("orgNo", serviceOwnerOrgNo);
-        const tokenGenerator = new EnterpriseTokenGenerator(tokenOpts);
-        serviceOwnerApiClient = new ServiceOwnerApiClient(__ENV.BASE_URL, tokenGenerator);
-    }
-    if (userApiClient == undefined) {
-        const tokenOpts = new Map();
-        tokenOpts.set("env", __ENV.ENVIRONMENT);
-        tokenOpts.set("ttl", 3600);
-        tokenOpts.set("scopes", "altinn:pdp/authorize.enduser");
-        personalTokenGenerator = new PersonalTokenGenerator(tokenOpts);
-        userApiClient = new BffUserApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        accessManagementApiClient = new BffAccessManagementApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        bffConnectionsApiClient = new BffConnectionsApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        bffAccessPackageApiClient = new BffAccessPackageApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        graphqlClient = new GraphqlClient(__ENV.BASE_URL, personalTokenGenerator);
-    }
-    return [serviceOwnerApiClient, userApiClient, accessManagementApiClient, bffConnectionsApiClient, bffAccessPackageApiClient, graphqlClient, personalTokenGenerator];
-}
-
-/**
- * Setup function to segment data for VUs.
- */
-export function setup() {
-    const numberOfVUs = getNumberOfVUs();
-    const res = http.get(`https://raw.githubusercontent.com/Altinn/altinn-platform-validation-tests/refs/heads/performance-instance-delegation/K6/testdata/authentication/orgs-in-${__ENV.ENVIRONMENT}-with-party-uuid-v2.csv`);
-    const segmentedData = segmentData(parseCsvData(res.body), numberOfVUs);
-    return segmentedData;
-}
-
-/**
  * Main function to test instance delegation from user to user. 
  * The test will create a dialog, then delegate it to another user, 
  * and check that the delegation is successful by calling the same endpoints as 
@@ -186,7 +127,7 @@ export function setup() {
  * (The groups are not used for anything else than to be able to see the flow of the test)
  */
 export default function (data) {
-    const [serviceOwnerApiClient, userApiClient, accessManagementApiClient, bffConnectionsApiClient, bffAccessPackageApiClient, graphqlClient, tokenGenerator] = getClients();
+    const [serviceOwnerApiClient, userApiClient, accessManagementApiClient, bffConnectionsApiClient, bffAccessPackageApiClient, graphqlClient, tokenGenerator] = getClients(serviceOwnerOrgNo);
     const { from, to } = getFromTo(data[exec.vu.idInTest - 1]);
     console.log(`VU: ${exec.vu.idInTest} - from: ${from.ssn} (${from.orgNo}) to: ${to.ssn}`);
     const resource = getItemFromList(resources);
@@ -212,45 +153,24 @@ export default function (data) {
     // Open access management after creating the dialog.
     // Call every bff endpoint that the browser uses when navigating from arbeidsflate/del og gi tilgang
     group(group1Label, function () {
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/lookup/party/user
         GetLookupPartyUser(userApiClient, getLookupPartyUserLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/isCompanyProfileAdmin?party=f1399084-2814-4f54-ab0e-75a931628762
         GetIsCompanyProfileAdmin(userApiClient, { party: from.orgUuid }, getIsCompanyProfileAdminLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/reportee/51267137
         GetReportee(userApiClient, from.partyId, getReporteeLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/profile
         GetProfile(userApiClient, getProfileLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/isAdmin?party=f1399084-2814-4f54-ab0e-75a931628762
         GetIsAdmin(userApiClient, { party: from.orgUuid }, getIsAdminLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/isClientAdmin?party=f1399084-2814-4f54-ab0e-75a931628762
         GetIsClientAdmin(userApiClient, { party: from.orgUuid }, getIsClientAdminLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/actorlist/old
         GetActorListOld(userApiClient, getActorListOldLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/actorlist/favorites
         GetActorListFavorites(userApiClient, getActorListFavoritesLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/role/permissions?party=5f453a8c-86e2-4bef-bbd9-6235edf414f0&from=f1399084-2814-4f54-ab0e-75a931628762&to=5f453a8c-86e2-4bef-bbd9-6235edf414f0
         GetRolePermissions(accessManagementApiClient, { party: from.partyUuid, from: from.orgUuid, to: from.partyUuid }, getRolePermissionsLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/cdn/orgdata
         GetOrganizationData(accessManagementApiClient, {}, getOrganizationDataLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/user/isInstanceAdmin?party=f1399084-2814-4f54-ab0e-75a931628762
         GetIsInstanceAdmin(userApiClient, { party: from.orgUuid }, getIsInstanceAdminLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/instances/delegation/instances?party=f1399084-2814-4f54-ab0e-75a931628762&from=f1399084-2814-4f54-ab0e-75a931628762&to=&resource=k6-instancedelegation-test&instance=urn%3Aaltinn%3Adialog-id%3A019d24e3-4c4b-72d7-97ba-eb55c92c4263
         GetDelegatedInstancesForResource(accessManagementApiClient, { party: from.partyUuid, from: from.partyUuid, resource: resource, instance: `urn:altinn:dialog-id:${dialogId}` }, getDelegatedInstancesForResourceLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/consent/active/f1399084-2814-4f54-ab0e-75a931628762
         GetActiveConsent(accessManagementApiClient, from.orgUuid, getActiveConsentLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/systemuser/f1399084-2814-4f54-ab0e-75a931628762/pending
         GetPendingDelegationsForUser(accessManagementApiClient, from.orgUuid, GetPendingDelegationsForUserLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/connection/rightholders?party=f1399084-2814-4f54-ab0e-75a931628762&from=f1399084-2814-4f54-ab0e-75a931628762&to=&includeClientDelegations=true&includeAgentConnections=true
         GetConnections(bffConnectionsApiClient, { party: from.orgUuid, from: from.orgUuid, includeClientDelegations: true, includeAgentConnections: true }, getConnectionsLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/resources?resourceId=k6-instancedelegation-test
         GetResourceById(accessManagementApiClient, { resourceId: resource }, getResourceByIdLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/accesspackage/delegationcheck?party=f1399084-2814-4f54-ab0e-75a931628762
         GetDelegationCheck(bffAccessPackageApiClient, { party: from.partyUuid }, getDelegationCheckLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/connection/rightholders?party=5f453a8c-86e2-4bef-bbd9-6235edf414f0&from=f1399084-2814-4f54-ab0e-75a931628762&to=5f453a8c-86e2-4bef-bbd9-6235edf414f0&includeClientDelegations=true&includeAgentConnections=true
         GetConnections(bffConnectionsApiClient, { party: from.partyUuid, from: from.partyUuid, to: from.partyUuid, includeClientDelegations: true, includeAgentConnections: true }, getConnectionsWithTo);
-
-
-
     });
 
     // // Delegate dialog to other user.
@@ -258,17 +178,11 @@ export default function (data) {
     group(group2Label, function () {
         const resp = GetRightsMeta(accessManagementApiClient, { resource: resource }, getRightsMetaLabel);
         CheckDelegationForResource(accessManagementApiClient, { party: from.orgUuid, resource: resource, instance: `urn:altinn:dialog-id:${dialogId}` }, checkDelegationForResourceLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/instances/delegation/instances/rights?party=f1399084-2814-4f54-ab0e-75a931628762&resource=k6-instancedelegation-test&instance=urn%3Aaltinn%3Adialog-id%3A019d24e3-4c4b-72d7-97ba-eb55c92c4263
         DelegateRightsForResource(accessManagementApiClient, { party: from.orgUuid, resource: resource, instance: `urn:altinn:dialog-id:${dialogId}` }, getInstanceDelegationBody(JSON.parse(resp), to), delegateRightsForResourceLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/instances/delegation/instances?party=f1399084-2814-4f54-ab0e-75a931628762&from=f1399084-2814-4f54-ab0e-75a931628762&to=&resource=k6-instancedelegation-test&instance=urn%3Aaltinn%3Adialog-id%3A019d24e3-4c4b-72d7-97ba-eb55c92c4263
         GetDelegatedInstancesForResource(accessManagementApiClient, { party: from.orgUuid, from: from.orgUuid, resource: resource, instance: `urn:altinn:dialog-id:${dialogId}` }, getDelegatedInstancesForResourceAfterLabel);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/instances/delegationcheck?party=f1399084-2814-4f54-ab0e-75a931628762&resource=k6-instancedelegation-test&instance=urn%3Aaltinn%3Adialog-id%3A019d24e3-4c4b-72d7-97ba-eb55c92c4263
         CheckDelegationForResource(accessManagementApiClient, { party: from.orgUuid, resource: resource, instance: `urn:altinn:dialog-id:${dialogId}` }, checkDelegationForResourceLabelAfter);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/connection/rightholders?party=5f453a8c-86e2-4bef-bbd9-6235edf414f0&from=f1399084-2814-4f54-ab0e-75a931628762&to=5f453a8c-86e2-4bef-bbd9-6235edf414f0&includeClientDelegations=true&includeAgentConnections=true
         GetConnections(bffConnectionsApiClient, { party: from.partyUuid, from: from.orgUuid, to: from.partyUuid, includeClientDelegations: true, includeAgentConnections: true }, getConnectionsWithToAfter);
-        //https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/connection/rightholders?party=f1399084-2814-4f54-ab0e-75a931628762&from=f1399084-2814-4f54-ab0e-75a931628762&to=&includeClientDelegations=true&includeAgentConnections=true
         GetConnections(bffConnectionsApiClient, { party: from.orgUuid, from: from.orgUuid, includeClientDelegations: true, includeAgentConnections: true }, getConnectionsLabelAfter);
-        //     GetRoleMeta(accessManagementApiClient, {}, getRoleMetaLabel);
     });
 
     group(group3Label, function () {
@@ -278,47 +192,3 @@ export default function (data) {
     });
 }
 
-function getDialogportenOpts(ssn) {
-    const tokenOpts = new Map();
-    tokenOpts.set("env", __ENV.ENVIRONMENT);
-    tokenOpts.set("ttl", 3600);
-    tokenOpts.set("scopes", "digdir:dialogporten");
-    tokenOpts.set("pid", ssn);
-    return tokenOpts;
-}
-
-/**
- * Helper function to create the body for delegating rights for a resource and instance to another user, 
- * based on the rights meta for the resource and the "to" user.
- * @param { JSON } rightsMeta 
- * @param {*} to 
- * @returns 
- */
-function getInstanceDelegationBody(rightsMeta, to) {
-    return {
-        to: {
-            personIdentifier: to.ssn,
-            lastName: to.lastName,
-        },
-        directRightKeys: rightsMeta.map((right) => right.key),
-    };
-}
-
-/**
- * Helper function to get from and to organizations/users for the current iteration, ensuring that they are not the same
- * @returns object with from and to organizations
- */
-function getFromTo(list) {
-    let from = undefined;
-    if (randomize) {
-        from = getItemFromList(list, randomize);
-    } else {
-        from = list[__ITER % list.length];
-    }
-    let to = getItemFromList(list, true);
-    while (to.ssn === from.ssn) {
-        to = getItemFromList(list, true);
-    }
-    console.log(`Selected from: ${from.ssn} and to: ${to.ssn}`);
-    return { from, to };
-}
