+++ K6/api/tests/authentication/access-management/bff/instance-delegation-user-user.js
@@ -1,10 +1,6 @@
-import http from "k6/http";
 import exec from "k6/execution";
 import { group } from "k6";
-import { EnterpriseTokenGenerator, PersonalTokenGenerator } from "../../../../../common-imports.js";
 import { CreateDialog } from "../../../../building-blocks/dialogporten/serviceowner/index.js";
-import { ServiceOwnerApiClient } from "../../../../../clients/dialogporten/serviceowner/index.js";
-import { GraphqlClient } from "../../../../../clients/dialogporten/graphql/index.js";
 import { GetAllDialogsForPartyCheckForDialogId, GetAndVerifyDialogById } from "../../../../building-blocks/dialogporten/graphql/index.js";
 import {
     GetLookupPartyUser,
@@ -18,14 +14,12 @@ import {
     GetIsInstanceAdmin,
 } from "../../../../building-blocks/authentication/instance-delegation/index.js";
 import { GetDelegatedInstancesForResource, GetActiveConsent, GetResourceById, GetRightsMeta, DelegateRightsForResource } from "../../../../building-blocks/authentication/client-delegations/index.js";
-import { BffUserApiClient, BffAccessManagementApiClient, BffConnectionsApiClient, BffAccessPackageApiClient } from "../../../../../clients/authentication/index.js";
 import { GetOrganizationData, CheckDelegationForResource, GetRoleMeta } from "../../../../building-blocks/authentication/client-delegations/index.js";
 import { GetConnections } from "../../../../building-blocks/authentication/connections/index.js";
 import { GetDelegationCheck } from "../../../../building-blocks/authentication/access-package/index.js";
-import { getItemFromList, parseCsvData, segmentData, getNumberOfVUs, getOptions } from "../../../../../helpers.js";
-import { getTokenOpts } from "./commons.js";
-
-const randomize = __ENV.RANDOMIZE ? __ENV.RANDOMIZE.toLowerCase() === "true" : false;
+import { getItemFromList, getOptions } from "../../../../../helpers.js";
+import { getTokenOpts, getFromTo, getClients, getDialogportenOpts, getInstanceDelegationBody } from "./commons.js";
+export { setup } from "./commons.js";
 
 // serviceowner which will create a dialog.
 // The yt serviceOwner is different from the other environments.
@@ -40,17 +34,6 @@ if (__ENV.ENVIRONMENT === "yt01") {
 const resources = [
     "k6-instancedelegation-test",
 ];
-
-// All apiclient used in this test
-let serviceOwnerApiClient = undefined;
-let userApiClient = undefined;
-let accessManagementApiClient = undefined;
-let bffConnectionsApiClient = undefined;
-let bffAccessPackageApiClient = undefined;
-let graphqlClient = undefined;
-
-// personal tokengenerator.
-let personalTokenGenerator = undefined;
 
 // Use unique laber for each request to be able to check them separately in the results.
 // Number them to make sumary report easier to read and to be able to see the flow of the test in the results. 
@@ -132,47 +115,6 @@ export const options = getOptions([
 );
 
 /**
-* Function to set up and return clients to interact with the Service Owner Dialog API
-*
-* @returns {Array} An array containing the AuthorizedPartiesClient instance
-*/
-export function getClients() {
-    if (serviceOwnerApiClient == undefined) {
-        const tokenOpts = new Map();
-        tokenOpts.set("env", __ENV.ENVIRONMENT);
-        tokenOpts.set("ttl", 3600);
-        tokenOpts.set("scopes", "digdir:dialogporten.serviceprovider");
-        tokenOpts.set("org", "ttd");
-        tokenOpts.set("orgNo", serviceOwnerOrgNo);
-        const tokenGenerator = new EnterpriseTokenGenerator(tokenOpts);
-        serviceOwnerApiClient = new ServiceOwnerApiClient(__ENV.BASE_URL, tokenGenerator);
-    }
-    if (userApiClient == undefined) {
-        const tokenOpts = new Map();
-        tokenOpts.set("env", __ENV.ENVIRONMENT);
-        tokenOpts.set("ttl", 3600);
-        tokenOpts.set("scopes", "altinn:pdp/authorize.enduser");
-        personalTokenGenerator = new PersonalTokenGenerator(tokenOpts);
-        userApiClient = new BffUserApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        accessManagementApiClient = new BffAccessManagementApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        bffConnectionsApiClient = new BffConnectionsApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        bffAccessPackageApiClient = new BffAccessPackageApiClient(__ENV.AM_UI_BASE_URL, personalTokenGenerator);
-        graphqlClient = new GraphqlClient(__ENV.BASE_URL, personalTokenGenerator);
-    }
-    return [serviceOwnerApiClient, userApiClient, accessManagementApiClient, bffConnectionsApiClient, bffAccessPackageApiClient, graphqlClient, personalTokenGenerator];
-}
-
-/**
- * Setup function to segment data for VUs.
- */
-export function setup() {
-    const numberOfVUs = getNumberOfVUs();
-    const res = http.get(`https://raw.githubusercontent.com/Altinn/altinn-platform-validation-tests/refs/heads/performance-instance-delegation/K6/testdata/authentication/orgs-in-${__ENV.ENVIRONMENT}-with-party-uuid-v2.csv`);
-    const segmentedData = segmentData(parseCsvData(res.body), numberOfVUs);
-    return segmentedData;
-}
-
-/**
  * Main function to test instance delegation from user to user. 
  * The test will create a dialog, then delegate it to another user, 
  * and check that the delegation is successful by calling the same endpoints as 
@@ -182,7 +124,7 @@ export function setup() {
  * (The groups are not used for anything else than to be able to see the flow of the test)
  */
 export default function (data) {
-    const [serviceOwnerApiClient, userApiClient, accessManagementApiClient, bffConnectionsApiClient, bffAccessPackageApiClient, graphqlClient, tokenGenerator] = getClients();
+    const [serviceOwnerApiClient, userApiClient, accessManagementApiClient, bffConnectionsApiClient, bffAccessPackageApiClient, graphqlClient, tokenGenerator] = getClients(serviceOwnerOrgNo);
     const { from, to } = getFromTo(data[exec.vu.idInTest - 1]);
     const resource = getItemFromList(resources);
     let dialogId = null;
@@ -243,53 +185,4 @@ export default function (data) {
         GetAllDialogsForPartyCheckForDialogId(graphqlClient, from.ssn, dialogId, getAllDialogsForPartyLabel);
         GetAndVerifyDialogById(graphqlClient, dialogId, getDialogByIdLabel);
     });
-
-    // Check these two endpoints, which returns 404 now. Called when opening access management
-    // https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/request/sent?party=5f453a8c-86e2-4bef-bbd9-6235edf414f0&status=Pending (404)
-    // https://am.ui.at23.altinn.cloud/accessmanagement/api/v1/request/received?party=5f453a8c-86e2-4bef-bbd9-6235edf414f0&status=Pending (404)
-}
-
-function getDialogportenOpts(ssn) {
-    const tokenOpts = new Map();
-    tokenOpts.set("env", __ENV.ENVIRONMENT);
-    tokenOpts.set("ttl", 3600);
-    tokenOpts.set("scopes", "digdir:dialogporten");
-    tokenOpts.set("pid", ssn);
-    return tokenOpts;
-}
-
-/**
- * Helper function to create the body for delegating rights for a resource and instance to another user, 
- * based on the rights meta for the resource and the "to" user.
- * @param { JSON } rightsMeta 
- * @param {*} to 
- * @returns 
- */
-function getInstanceDelegationBody(rightsMeta, to) {
-    return {
-        to: {
-            personIdentifier: to.ssn,
-            lastName: to.lastName,
-        },
-        directRightKeys: rightsMeta.map((right) => right.key),
-    };
-}
-
-/**
- * Helper function to get from and to organizations/users for the current iteration, ensuring that they are not the same
- * @returns object with from and to organizations
- */
-function getFromTo(list) {
-    let from = undefined;
-    if (randomize) {
-        from = getItemFromList(list, randomize);
-    } else {
-        from = list[__ITER % list.length];
-    }
-    let to = getItemFromList(list, true);
-    while (to.ssn === from.ssn) {
-        to = getItemFromList(list, true);
-    }
-    console.log(`Selected from: ${from.ssn} and to: ${to.ssn}`);
-    return { from, to };
 }
